Split the cache-step anatomy out of workflow_support
`tests/workflow_contracts/workflow_support.py` was the one module in the
tree over the 400-line budget, at 501 lines, and managed PyPy cannot
parse it: it uses PEP 758's parenthesis-free `except A, B:`, which CPython
3.14 accepts and the lagging grammar rejects. PyLint therefore reports
nothing at all for the module, not even `too-many-lines`, so no gate
noticed. Splitting it is what makes the line budget enforceable on it.

The cache-step cluster moves wholesale: the approved action, its single
pinned ref, the workspace test drivers, and the five predicates built on
them now live in `cache_step_support.py`, 142 lines. The loaders,
constants, and error classes stay behind, so no exception identity
changes and the module needs no new imports -- only its own cache
constants and predicates leave. `workflow_queries.py` and
`runner_cache_test.py` are the two importers that follow.

No behaviour changes; the moved code is byte-for-byte the code that left.
`make test-workflow-contracts` passes at 166.

Refs #742.
