+++ tests/workflow_contracts/workflow_queries.py
@@ -12,17 +12,15 @@ import shutil
 import subprocess  # ruff: ignore[suspicious-subprocess-import] - runs a script this repository declares.
 import typing as typ
 
-from publish_report_support import verification_script
-from workflow_support import (
-    ROOT,
+from cache_step_support import (
     cache_owner,
     cache_paths,
     is_cache_step,
-    jobs,
     path_components,
     runs_workspace_tests,
-    steps,
 )
+from publish_report_support import verification_script
+from workflow_support import ROOT, jobs, steps
 
 if typ.TYPE_CHECKING:
     import collections.abc as cabc
